refactor(web): move ChatGptSignIn to its own file
It sat inline in a 533-line panel while its Google counterpart already
had a file, and the panel is about to gain a provider-first layout.

Pure move: the component and the loginError helper it uses are unchanged.
